+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/index/BTreeIndexFileReader.scala
@@ -22,86 +22,27 @@ import org.apache.hadoop.fs.Path
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.execution.datasources.OapException
-import org.apache.spark.sql.execution.datasources.oap.filecache.{FiberCache, MemoryManager}
 import org.apache.spark.sql.execution.datasources.oap.io.IndexFile
-import org.apache.spark.sql.internal.oap.OapConf
-import org.apache.spark.unsafe.Platform
-import org.apache.spark.util.ShutdownHookManager
 
-private[oap] case class BTreeIndexFileReader(
-    configuration: Configuration,
-    file: Path) extends Logging {
 
-  private val VERSION_SIZE = IndexFile.VERSION_LENGTH
-  private val FOOTER_LENGTH_SIZE = IndexUtils.INT_SIZE
-  private val ROW_ID_LIST_LENGTH_SIZE = IndexUtils.LONG_SIZE
+trait BTreeIndexFileReader extends Logging
+object BTreeIndexFileReader {
 
-  // Section ID for fiber cache reading.
-  val footerSectionId: Int = 0
-  val rowIdListSectionId: Int = 1
-  val nodeSectionId: Int = 2
+  private val MAGIC_V1 = IndexUtils.serializeVersion(1)
 
-  val rowIdListSizePerSection: Int =
-    configuration.getInt(OapConf.OAP_BTREE_ROW_LIST_PART_SIZE.key, 1024 * 1024)
-
-  private lazy val (reader, fileLength) = {
+  def apply(configuration: Configuration, file: Path): BTreeIndexFileReader = {
     val fs = file.getFileSystem(configuration)
-    (fs.open(file), fs.getFileStatus(file).getLen)
-  }
+    val reader = fs.open(file)
+    val fileLen = fs.getFileStatus(file).getLen
 
-  private val (footerLength, rowIdListLength) = {
-    val sectionLengthIndex = fileLength - FOOTER_LENGTH_SIZE - ROW_ID_LIST_LENGTH_SIZE
-    val sectionLengthBuffer = new Array[Byte](FOOTER_LENGTH_SIZE + ROW_ID_LIST_LENGTH_SIZE)
-    reader.readFully(sectionLengthIndex, sectionLengthBuffer)
-    val rowIdListSize = getLongFromBuffer(sectionLengthBuffer, 0)
-    val footerSize = getIntFromBuffer(sectionLengthBuffer, ROW_ID_LIST_LENGTH_SIZE)
-    (footerSize, rowIdListSize)
-  }
+    val magic = new Array[Byte](IndexFile.VERSION_LENGTH)
 
-  private def footerIndex = fileLength - FOOTER_LENGTH_SIZE - ROW_ID_LIST_LENGTH_SIZE - footerLength
-  private def rowIdListIndex = footerIndex - rowIdListLength
-  private def nodesIndex = VERSION_SIZE
-
-  private def getLongFromBuffer(buffer: Array[Byte], offset: Int) =
-    Platform.getLong(buffer, Platform.BYTE_ARRAY_OFFSET + offset)
-
-  private def getIntFromBuffer(buffer: Array[Byte], offset: Int) =
-    Platform.getInt(buffer, Platform.BYTE_ARRAY_OFFSET + offset)
-
-  def checkVersionNum(versionNum: Int): Unit = {
-    if (IndexFile.VERSION_NUM != versionNum) {
-      throw new OapException("Btree Index File version is not compatible!")
-    }
-  }
-
-  def readFooter(): FiberCache =
-    MemoryManager.putToIndexFiberCache(reader, footerIndex, footerLength)
-
-  def readRowIdList(partIdx: Int): FiberCache = {
-    val partSize = rowIdListSizePerSection.toLong * IndexUtils.INT_SIZE
-    val readLength = if (partIdx * partSize + partSize > rowIdListLength) {
-      rowIdListLength % partSize
+    reader.readFully(0, magic)
+    if (magic.sameElements(MAGIC_V1)) {
+      BTreeIndexFileReaderV1(configuration, reader, file, fileLen)
     } else {
-      partSize
+      throw new OapException("not a valid index file")
     }
-    assert(readLength <= Int.MaxValue, "Size of each row id list partition is too large!")
-    MemoryManager.putToIndexFiberCache(reader, rowIdListIndex + partIdx * partSize,
-      readLength.toInt)
   }
 
-  @deprecated("no need to read the whole row id list", "v0.3")
-  def readRowIdList(): FiberCache =
-    MemoryManager.putToIndexFiberCache(reader, rowIdListIndex, rowIdListLength.toInt)
-
-  def readNode(offset: Int, size: Int): FiberCache =
-    MemoryManager.putToIndexFiberCache(reader, nodesIndex + offset, size)
-
-  def close(): Unit = try {
-    reader.close()
-  } catch {
-    case e: Exception =>
-      if (!ShutdownHookManager.inShutdown()) {
-        logWarning("Exception in FSDataInputStream.close()", e)
-      }
-  }
 }
